CubeModel uses byte index

+++ Demos/CSharpGL.Demos/Models/CubeModel.cs
@@ -52,8 +52,8 @@ namespace CSharpGL.Models
             faceNZ = new SqureColor(new vec3(1, 1, 0)),
         };
 
-        static readonly uint[] identityIndex =
-            new uint[] { 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, 8, 9, 10, 8, 10, 11, 12, 13, 14, 12, 14, 15, 16, 17, 18, 16, 18, 19, 20, 21, 22, 20, 22, 23 };
+        static readonly byte[] identityIndex =
+            new byte[] { 0, 1, 2, 0, 2, 3, 4, 5, 6, 4, 6, 7, 8, 9, 10, 8, 10, 11, 12, 13, 14, 12, 14, 15, 16, 17, 18, 16, 18, 19, 20, 21, 22, 20, 22, 23 };
 
         static CubeColor GetRandomColor()
         {
@@ -72,7 +72,7 @@ namespace CSharpGL.Models
         public CubePosition position;
         public CubeNormal normal;
         public CubeColor color;
-        public uint[] index;
+        public byte[] index;
 
         public CubeModel(float radius)
         {
